serve_manager/requirements.txt
fastapi
uvicorn
jinja2
psutil
psycopg2-binary
python-dotenv
sqlalchemy